refactor(serve): extract PromQL evaluator into dedicated module
Move parse_metrics_text, match_metric, find_top_level_binary_op,
binary_op, and eval_instant_query into inference_pipeline/promql.py.
The evaluator now takes a metrics_text_fn callable, decoupling it
from the serve module's metric aggregation concern.

serve.py: 516 → 288 lines.
